#FIX View클래스들 자원화, Swapchain 분할중 (RTV/DSV/SRV)

+++ C_CPP/PardCode29/CommonHeader.h
@@ -111,8 +111,8 @@ struct PointXY
 
 struct RenderCounts
 {
-	size_t count;
-	size_t idx;
+	UINT count;
+	UINT idx;
 };
 
 struct Vertex_PC
